Refactor SQL extraction functions in TypeScript and Rust
