+++ config/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "outDir": "out",
+    "rootDir": "src",
+    "esModuleInterop": false,
+    "declaration": true
+  },
+  "include": ["src/**/*"],
+  "extends": ["../tsconfig.json"]
+}
