+++ pnpm-workspace.yaml
@@ -1,3 +1,4 @@
 packages:
-  - vsce
+  - config
   - sql-extraction/*
+  - vsce
