+++ sql-extraction/rs/src/index.ts
@@ -1,17 +1,4 @@
-// TODO: DI
-type SqlNode = {
-  code_range: {
-    start: {
-      line: number; // 0-based
-      character: number; // 0-based
-    };
-    end: {
-      line: number; // 0-based
-      character: number; // 0-based
-    };
-  };
-  content: string;
-};
+import type { SqlNode } from "@senken/config";
 
 export async function extractSqlListRs(sourceTxt: string): Promise<SqlNode[]> {
   const { extract_sql_list } = await import("../pkg");
